If gun is cold for too long, swith to other gun
If my first gun linear or circular, oscillatory bot tricks it to large
bacj and force swing, as result gun is never settles and never fires.
This patch crudly fix this situation, by picking (probabilisticly)
another gun.

Make sure that either random or some brainless gun is in the toolbox!